+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Added
+- Add support for the `in_app_include` and `in_app_exclude` options
+
 ### Fixed
 - Replace recommended guzzlehttp/psr7 with php-http/guzzle6-adapter